Update u3lab.md
Stayed up late last night with Het and built/worked/completed the first version of this lab.

This lab is confusing as *%^& to begin with.

I used Het's final pdf as a base and then added prompts before commands where they were missing.

This is mostly work on the LDAP portion on the bottom.  

It is not formatted perfectly and needs a lot more work.  

This commit is just to capture the first run through while the lab was complete initially.